+++ src/main/kotlin/com/nylas/models/ListFoldersQueryParams.kt
@@ -21,13 +21,21 @@ data class ListFoldersQueryParams(
   /**
    * (Microsoft and EWS only.) Use the ID of a folder to find all child folders it contains.
    */
-  @Json(name = " parent_id")
+  @Json(name = "parent_id")
   val parentId: String? = null,
+  /**
+   * Specify fields that you want Nylas to return, as a comma-separated list (for example, select=id,updated_at).
+   * This allows you to receive only the portion of object data that you're interested in.
+   * You can use select to optimize response size and reduce latency by limiting queries to only the information that you need
+   */
+  @Json(name = "select")
+  var select: String? = null,
 ) : IQueryParams {
   class Builder {
     private var limit: Int? = null
     private var pageToken: String? = null
     private var parentId: String? = null
+    private var select: String? = null
 
     /**
      * Sets the maximum number of objects to return.
@@ -53,6 +61,13 @@ data class ListFoldersQueryParams(
     fun parentId(parentId: String?) = apply { this.parentId = parentId }
 
     /**
+     * Sets the fields to return in the response.
+     * @param select List of field names to return (e.g. "id,updated_at")
+     * @return The builder.
+     */
+    fun select(select: String?) = apply { this.select = select }
+
+    /**
      * Builds the [ListFoldersQueryParams] object.
      * @return The [ListFoldersQueryParams] object.
      */
@@ -60,6 +75,7 @@ data class ListFoldersQueryParams(
       limit = limit,
       pageToken = pageToken,
       parentId = parentId,
+      select = select,
     )
   }
 }